Move types file to a folder structure
