+++ apps/ledger-live-desktop/src/renderer/families/evm/types.ts
@@ -1,4 +1,4 @@
-import { Transaction, TransactionStatus } from "@ledgerhq/coin-evm/types";
+import { Transaction, TransactionStatus } from "@ledgerhq/coin-evm/types/index";
 
 import { Account } from "@ledgerhq/types-live";
 import { LLDCoinFamily } from "../types";
