+++ libs/coin-evm/src/bridge/js.ts
@@ -5,9 +5,9 @@ import {
   makeScanAccounts,
 } from "@ledgerhq/coin-framework/bridge/jsHelpers";
 import { SignerContext } from "@ledgerhq/coin-framework/signer";
-import type { AccountBridge, CurrencyBridge } from "@ledgerhq/types-live";
+import type { AccountBridge, Bridge, CurrencyBridge } from "@ledgerhq/types-live";
 import getAddressWrapper from "@ledgerhq/coin-framework/bridge/getAddressWrapper";
-import type { Transaction as EvmTransaction } from "../types";
+import type { Transaction as EvmTransaction } from "../types/index";
 import { EvmAddress, EvmSignature, EvmSigner } from "../signer";
 import { estimateMaxSpendable } from "../estimateMaxSpendable";
 import { getTransactionStatus } from "../getTransactionStatus";
@@ -59,7 +59,9 @@ export function buildAccountBridge(
   };
 }
 
-export function createBridges(signerContext: SignerContext<EvmSigner, EvmAddress | EvmSignature>) {
+export function createBridges(
+  signerContext: SignerContext<EvmSigner, EvmAddress | EvmSignature>,
+): Bridge<EvmTransaction> {
   return {
     currencyBridge: buildCurrencyBridge(signerContext),
     accountBridge: buildAccountBridge(signerContext),
